+++ apps/nowait-admin/src/hooks/booth/menu/useDeleteMenu.tsx
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query";
-import AdminApi from "../../utils/AdminApi";
+import AdminApi from "../../../utils/AdminApi";
 
 export const useDeleteMenu = () => {
   return useMutation({
