+++ apps/nowait-admin/src/hooks/booth/menu/useGetAllMenus.tsx
@@ -1,26 +1,32 @@
 import { useQuery } from "@tanstack/react-query";
-import AdminApi from "../../utils/AdminApi";
+import AdminApi from "../../../utils/AdminApi";
 
-interface MenuItem {
+interface MenuImage {
+  id: number;
+  imageUrl: string;
+}
+
+interface RawMenuItem {
   menuId: number;
+  storeId: number;
   name: string;
   description: string;
   price: number;
   isSoldOut: boolean;
-  imageUrl?: string;
-  // 필요한 경우 필드를 더 추가하세요
+  deleted: boolean;
+  images: MenuImage[];
 }
 
 export const useGetAllMenus = (storeId: number) => {
-  return useQuery<MenuItem[]>({
+  return useQuery<RawMenuItem[]>({
     queryKey: ["all-menus", storeId],
     queryFn: async () => {
       const res = await AdminApi.get(
         `/admin/menus/all-menus/stores/${storeId}`
       );
-      return res.data;
+      return res.data.response.menuReadDto; // ✅ 여기만 추출
     },
     enabled: !!storeId,
-    staleTime: 1000 * 60, // 1분 캐싱
+    staleTime: 1000 * 60,
   });
 };
