+++ apps/nowait-admin/src/pages/AdminBooth/AdminBooth.tsx
@@ -5,12 +5,7 @@ import placeholderIcon from "../../assets/image_placeholder.svg";
 import { useWindowWidth } from "../../hooks/useWindowWidth";
 import OperatingTimeSelector from "./components/OperatingTimeSelector";
 import NoticeEditor from "./components/NoticeEditor";
-import editOrderIcon from "../../assets/edit_order_icon.svg";
-import ToggleSwitch from "../AdminHome/components/ToggleSwitch";
-import { DragDropContext, Droppable, Draggable } from "react-beautiful-dnd";
-import MenuModal from "./components/menuModal";
-import { useCreateMenu } from "../../hooks/booth/useCreateMenu";
-import { useUploadMenuImage } from "../../hooks/booth/useUploadMenuImage";
+import MenuSection from "./components/MenuSection";
 
 const BoothSection = ({
   boothName,
@@ -149,207 +144,6 @@ const BoothSection = ({
     </div>
   </>
 );
-
-const dummyMenus = [
-  { name: "갈비구이", price: "10,000원", soldOut: false },
-  { name: "메뉴명", price: "가격", soldOut: false },
-  { name: "메뉴명", price: "가격", soldOut: false },
-  { name: "메뉴명", price: "가격", soldOut: false },
-  { name: "메뉴명", price: "가격", soldOut: false },
-  { name: "메뉴명", price: "가격", soldOut: false },
-  { name: "메뉴명", price: "가격", soldOut: false },
-  { name: "메뉴명", price: "가격", soldOut: false },
-  { name: "메뉴명", price: "가격", soldOut: false },
-];
-
-const MenuSection = () => {
-  const [editMode, setEditMode] = useState(false);
-  const [menus, setMenus] = useState(dummyMenus);
-  const [isAddModalOpen, setIsAddModalOpen] = useState(false);
-  const [isEditModalOpen, setIsEditModalOpen] = useState(false);
-  const [selectedMenu, setSelectedMenu] = useState<any>(null);
-  const { mutate: createMenu } = useCreateMenu();
-  const { mutate: uploadMenuImage } = useUploadMenuImage();
-
-  const openEditModal = (menu: any) => {
-    setSelectedMenu(menu);
-    setIsEditModalOpen(true);
-  };
-
-  const handleAddMenu = (newMenu: any) => {
-    const payload = {
-      storeId: 1, // 추후 props 등으로 받아야 함
-      name: newMenu.name,
-      description: newMenu.description,
-      price: parseInt(newMenu.price.replace(/[^0-9]/g, ""), 10),
-    };
-
-    createMenu(payload, {
-      onSuccess: (res) => {
-        const createdMenuId = res.menuId; // 백엔드가 반환한 menuId 사용
-
-        // 이미지가 있을 경우 업로드
-        if (newMenu.image) {
-          uploadMenuImage(
-            { menuId: createdMenuId, image: newMenu.image },
-            {
-              onSuccess: () => {
-                alert("메뉴 및 이미지가 성공적으로 추가되었습니다.");
-                setMenus((prev) => [...prev, newMenu]);
-              },
-              onError: () => {
-                alert("메뉴는 추가되었지만 이미지 업로드에 실패했습니다.");
-              },
-            }
-          );
-        } else {
-          alert("메뉴가 성공적으로 추가되었습니다.");
-          setMenus((prev) => [...prev, newMenu]);
-        }
-
-        setIsAddModalOpen(false);
-      },
-      onError: () => {
-        alert("메뉴 추가에 실패했습니다.");
-      },
-    });
-  };
-
-  const handleEditMenu = (updated: any) => {
-    setMenus((prev) =>
-      prev.map((menu) => (menu.name === selectedMenu.name ? updated : menu))
-    );
-  };
-
-  const toggleSoldOut = (index: number) => {
-    const updatedMenus = [...menus];
-    updatedMenus[index].soldOut = !updatedMenus[index].soldOut;
-    setMenus(updatedMenus);
-  };
-
-  const handleDragEnd = (result: any) => {
-    if (!result.destination) return;
-
-    const reordered = Array.from(menus);
-    const [removed] = reordered.splice(result.source.index, 1);
-    reordered.splice(result.destination.index, 0, removed);
-    setMenus(reordered);
-  };
-
-  return (
-    <div className="mt-[40px] mb-[20px]">
-      <div className="flex justify-between items-center mb-[20px]">
-        <h2 className="text-title-18-bold">메뉴</h2>
-        <div className="flex gap-[10px]">
-          <button
-            className="text-sm px-4 py-2 bg-black-5 text-black-80 rounded-[8px]"
-            onClick={() => setEditMode((prev) => !prev)}
-          >
-            {editMode ? "편집 완료" : "순서 편집"}
-          </button>
-          <button
-            className="text-sm px-4 py-2 bg-black-5 text-black-80 rounded-[8px]"
-            onClick={() => setIsAddModalOpen(true)}
-          >
-            메뉴 추가 +
-          </button>
-        </div>
-      </div>
-
-      <div className="flex justify-between mb-[10px]">
-        <p className="text-sm text-black-40 mb-2">{menus.length}개의 메뉴</p>
-        {!editMode && <p className="text-sm text-black-40 mb-2">품절 표시</p>}
-      </div>
-
-      <div className="border-t border-[#EEEEEE]">
-        <DragDropContext onDragEnd={handleDragEnd}>
-          <Droppable
-            droppableId="menu-list"
-            isDropDisabled={!editMode}
-            direction="vertical"
-            // isCombineEnabled={false}
-            // ignoreContainerClipping={false}
-          >
-            {(provided) => (
-              <div {...provided.droppableProps} ref={provided.innerRef}>
-                {menus.map((menu, idx) => (
-                  <Draggable
-                    key={`menu-${idx}`}
-                    draggableId={`menu-${idx}`}
-                    index={idx}
-                    isDragDisabled={!editMode}
-                  >
-                    {(provided) => (
-                      <div
-                        className="flex justify-between items-center py-4"
-                        ref={provided.innerRef}
-                        {...provided.draggableProps}
-                        {...(editMode ? provided.dragHandleProps : {})}
-                      >
-                        <div className="flex items-center gap-4">
-                          <div
-                            className="w-[48px] h-[48px] bg-black-5 rounded-md flex items-center justify-center overflow-hidden"
-                            onClick={() => !editMode && openEditModal(menu)}
-                          >
-                            <img
-                              src={placeholderIcon}
-                              className="w-6 h-6"
-                              alt="placeholder"
-                            />
-                          </div>
-                          <div className="flex flex-col">
-                            <span className="text-sm font-semibold">
-                              {menu.name}
-                            </span>
-                            <span className="text-sm text-black-60">
-                              {menu.price}
-                            </span>
-                          </div>
-                        </div>
-
-                        <div className="text-sm text-black-60">
-                          {editMode ? (
-                            <img
-                              src={editOrderIcon}
-                              alt="순서 변경"
-                              className="w-5 h-5"
-                              {...provided.dragHandleProps}
-                            />
-                          ) : (
-                            <ToggleSwitch
-                              isOn={menu.soldOut}
-                              toggle={() => toggleSoldOut(idx)}
-                            />
-                          )}
-                        </div>
-                      </div>
-                    )}
-                  </Draggable>
-                ))}
-                {provided.placeholder}
-              </div>
-            )}
-          </Droppable>
-        </DragDropContext>
-      </div>
-      {isAddModalOpen && (
-        <MenuModal
-          isEdit={false}
-          onClose={() => setIsAddModalOpen(false)}
-          onSubmit={handleAddMenu}
-        />
-      )}
-      {isEditModalOpen && selectedMenu && (
-        <MenuModal
-          isEdit={true}
-          initialData={selectedMenu}
-          onClose={() => setIsEditModalOpen(false)}
-          onSubmit={handleEditMenu}
-        />
-      )}
-    </div>
-  );
-};
 
 const BoothForm = () => {
   const width = useWindowWidth();
